refactor: change quantity type from Long to BigDecimal across order and fill entities

+++ commons/src/test/java/com/pbkour/mintrade/commons/kafka/FillTest.java
@@ -26,7 +26,7 @@ class FillTest {
 
         Fill fill = Fill.builder()
             .fillId(id)
-            .quantity(100L)
+            .quantity(BigDecimal.valueOf(100))
             .price(new BigDecimal("123.45"))
             .timestamp(ts)
             .build();
